requirements.txt
# Dépendances pour l'application Gare de Train
# Installez avec: pip install -r requirements.txt

# Framework web principal
Flask==3.1.2

# Extensions Flask
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
Flask-Login==0.6.3

# Gestion des formulaires
WTForms==3.2.1

# Base de données
SQLAlchemy==2.0.43
psycopg2-binary==2.9.10

# Dépendances système
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.2
itsdangerous==2.2.0
click==8.3.0
blinker==1.9.0

# Développement (optionnel)
python-dotenv==1.0.0
